Moved methods into PayloadChecksHelper
