+++ api/KS.FiksProtokollValidator.WebAPI/Validation/FiksResponseValidator.cs
@@ -159,14 +159,14 @@ namespace KS.FiksProtokollValidator.WebAPI.Validation
             {
                 var xmlContent = System.Text.Encoding.Default.GetString(fiksPayload.Payload);
 
-                ValidateXmlWithSchema(xmlContent, validationErrors, messageType);
+                PayloadChecksHelper.ValidateXmlWithSchema(xmlContent, validationErrors, messageType);
                 ValidateXmlPayloadContent(xmlContent, fiksRequest, validationErrors);
             }
             else
             {
                 if (receivedPayloadFileName != null && receivedPayloadFileName.EndsWith(".json"))
                 {
-                    ValidateJsonWithSchema(System.Text.Encoding.Default.GetString(fiksPayload.Payload), validationErrors, messageType);
+                    PayloadChecksHelper.ValidateJsonWithSchema(System.Text.Encoding.Default.GetString(fiksPayload.Payload), validationErrors, messageType);
                     ValidateJsonPayloadContent(System.Text.Encoding.Default.GetString(fiksPayload.Payload), fiksRequest.TestCase.FiksResponseTests, validationErrors);
                 }
             }
@@ -186,64 +186,6 @@ namespace KS.FiksProtokollValidator.WebAPI.Validation
         private static bool HasCorrectFilename(string messageType, string filename)
         {
             return PayloadChecksHelper.GetExpectedFileName(messageType).Equals(filename);
-        }
-
-        private static void ValidateJsonWithSchema(string payload, List<string> validationErrors, string messageType)
-        {
-            var baseDirectory = Path.GetDirectoryName(Assembly.GetEntryAssembly()?.Location);
-            var pathToSchema = Path.Combine(baseDirectory, "Schema", messageType + ".schema.json");
-            using (TextReader file = File.OpenText(pathToSchema))
-            {
-                JObject jObject = JObject.Parse(payload);
-                JSchema schema = JSchema.Parse(file.ReadToEnd());
-
-                schema.ExtensionData.Remove("definitions");
-                AddAdditionalPropertiesFalseToSchemaProperties(schema.Properties);
-                schema.AllowAdditionalProperties = false;
-
-                //TODO:Skille mellom errors og warnings hvis det er 
-                jObject.Validate(schema, (o, a) =>
-                {
-                    validationErrors.Add(a.Message);
-                });
-            }
-        }
-
-        private static void AddAdditionalPropertiesFalseToSchemaProperties(IDictionary<string, JSchema> properties)
-        {
-            foreach (var item in properties)
-            {
-                item.Value.AllowAdditionalProperties = false;
-                foreach (var itemItem in item.Value.Items)
-                {
-                    AddAdditionalPropertiesFalseToSchemaProperties(itemItem.Properties);
-
-                }
-                AddAdditionalPropertiesFalseToSchemaProperties(item.Value.Properties);
-            }
-        }
-
-        private static void ValidateXmlWithSchema(string xmlPayloadContent, List<string> validationErrors, string messageType)
-        {
-            var xsdValidator = new XsdValidator();
-            switch (messageType)
-            {
-                case ArkivintegrasjonMeldingTypeV1.ArkivmeldingKvittering:
-                    xsdValidator.ValidateArkivmeldingKvittering(xmlPayloadContent, validationErrors);
-                    break;
-                case ArkivintegrasjonMeldingTypeV1.SokResultatMinimum:
-                   xsdValidator.ValidateArkivmeldingSokeresultatMinimum(xmlPayloadContent, validationErrors);
-                   break;
-                case ArkivintegrasjonMeldingTypeV1.SokResultatNoekler:
-                    xsdValidator.ValidateArkivmeldingSokeresultatNoekler(xmlPayloadContent, validationErrors);
-                    break;
-                case ArkivintegrasjonMeldingTypeV1.SokResultatUtvidet:
-                    xsdValidator.ValidateArkivmeldingSokeresultatUtvidet(xmlPayloadContent, validationErrors);
-                    break;
-                default:
-                    //do nothing? Or display a warning that the message type was not checked against xsd?
-                    break;
-            }
         }
 
        public static void ValidateXmlPayloadContent(string xmlPayloadContent, FiksRequest fiksRequest,
